requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
yt-dlp==2023.12.30
supabase==2.3.0
boto3==1.34.0
celery==5.3.4
redis==5.0.1
pydantic==2.5.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.0
python-dotenv==1.0.0
requests==2.31.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9